Minimum time-period site selection

This function uses part of the method outlined in Roy et al (2012) and Isaac et al (2014) for selecting well-sampled sites from a dataset using the number of time periods only. siteSelection is a wrapper for this function that performs the complete site selection process as outlined in these papers.

siteSelectionMinTP(taxa, site, time_period, minTP)

Arguments

taxa
A character vector of taxon names

site
A character vector of site names

time_period
A numeric vector of user defined time periods, or a date vector

minTP
numeric, The minimum number of time periods, or if time_period is a date the minimum number of years, a site must be sampled in for it be be considered well sampled.

Value

A data.frame of data that forefills the selection criteria. This data has two attributes: sites gives the total number of sites in the dataset, and sucess gives the number of sites that satify the selection criteria

References

needed