sparta Trend Analysis for Unstructured Data

The Sparta package includes methods used to analyse trends in unstructured occurrence datasets. Methods included in the package include Frescalo, Telfer's change index, Reporting rate models and Bayesian Occupancy models. These methods are reviewed in Issac et al (2014), available at http://onlinelibrary.wiley.com/doi/10.1111/2041-210X.12254/abstract

See also

Author

Maintainer: Dylan Carbone dylcar@ceh.ac.uk (ORCID)

Authors:

Nick Isaac (ORCID)

Tom August tomaug@ceh.ac.uk (ORCID)

Colin Harrower (ORCID)

Jack Hatfield (ORCID)

Mark Hill (ORCID)

Francesca Mancini (ORCID)

Charlie Outhwaite (ORCID)

Gary Powney (ORCID)