Format data for Occupancy detection models

This takes occurrene data in the form of a vector of taxa names, locations and time_period (usually a date) and converts them into the form needed for occupancy models (see value section)

formatOccData(taxa, site, time_period, includeJDay = FALSE)

Arguments

taxa
A character vector of taxon names, as long as the number of observations.

site
A character vector of site names, as long as the number of observations.

time_period
A numeric vector of user defined time periods, or a date vector, as long as the number of observations.

includeJDay
Logical. If TRUE a Julian day column is returned in the occDetData object

Value

A list of length 2 the first element 'spp_vis' is a data.frame with visit (unique combination of site and time period) in the first column and taxa for all the following columns. Values in taxa columns are either TRUE or FALSE depending on whether they were observed on that visit. The second element ('occDetData') is a dataframe giving the site, list length (the number of species observed on a visit) and year for each visit. Optionally this also includes a Julian Day column

References

Isaac, N.J.B., van Strien, A.J., August, T.A., de Zeeuw, M.P. and Roy, D.B. (2014). Statistics for citizen science: extracting signals of change from noisy ecological data. Methods in Ecology and Evolution, 5 (10), 1052-1060.

Examples

# NOT RUN {
# Create data
n <- 15000 #size of dataset
nyr <- 20 # number of years in data
nSamples <- 100 # set number of dates
nSites <- 50 # set number of sites

# Create somes dates
first <- as.Date(strptime("2010/01/01", "%Y/%m/%d"))
last <- as.Date(strptime(paste(2010+(nyr-1),"/12/31", sep=''), "%Y/%m/%d"))
dt <- last-first
rDates <- first + (runif(nSamples)*dt)

# taxa are set as random letters
taxa <- sample(letters, size = n, TRUE)

# three sites are visited randomly
site <- sample(paste('A', 1:nSites, sep=''), size = n, TRUE)

# the date of visit is selected at random from those created earlier
time_period <- sample(rDates, size = n, TRUE)

# run the model with these data for one species
formatted_data <- formatOccData(taxa = taxa,
                                site = site,
                                time_period = time_period)
# }