Telfer's change index

Telfers change index is designed to assess the relative change in range size of species between two time periods (Telfer et al 2002). This function can take multiple time periods and will complete all pairwise comparisons.

telfer(taxa, site, time_period, minSite = 5, useIterations = TRUE,
  iterations = 10)

Arguments

taxa
A character vector of taxon names

site
A character vector of site names

time_period
A numeric vector of user defined time periods, or a date vector

minSite
The minimum number of sites occupied in the first time period in order for a trend to be calculated for a taxon.

useIterations
A logical variable indicating whether iterations should be used. Iterations are used to account for increased variation in the logit proportions close to zero and one (see Telfer et al 2002). Default is TRUE

iterations
If useIterations is TRUE, then this parameter indicates the number of iterations to be used. In Telfer et al 2002 the number of iterations used were 7 and 8 for the two datasets for which it was applied. The defualt here is 10.

References

Telfer, M.G., Preston, C.D., & Rothery, P. (2002) A general method for measuring relative change in range size from biological atlas data. Biological Conservation, 107, 99-109.

Examples

# Create fake data
SS <- 5000 # number of observations
taxa <- sample(letters, SS, replace = TRUE)
site <- sample(paste('A', 1:20, sep = ''), SS, replace = TRUE)
time_period <- sample(1:3, SS, replace = TRUE)

TelferResult <- telfer(taxa, site, time_period)

#> Warning: 3496 out of 5000 observations will be removed as duplicates

head(TelferResult)

#>   taxa Nsite_1.x Nsite_2.x Telfer_1_2 Nsite_1.y Nsite_3.x Telfer_1_3 Nsite_2.y
#> 1    a        17        20  0.4178907        17        20  0.9430450        20
#> 2    b        19        20  0.7447182        19        18 -1.5673369        20
#> 3    c        19        18 -1.8782816        19        19 -0.7439658        18
#> 4    d        19        20  0.7447182        19        19 -0.7439658        20
#> 5    e        20        20  1.1633923        20        16 -1.9417249        20
#> 6    f        19        20  0.7447182        19        20  0.9344623        20
#>   Nsite_3.y Telfer_2_3
#> 1        20  1.1440416
#> 2        18 -1.0763359
#> 3        19 -1.1171519
#> 4        19 -0.3455839
#> 5        16 -1.9870010
#> 6        20  1.1440416